+++ src/main/scala/com/nec/cache/CachedVeBatch.scala
@@ -1,16 +1,16 @@
 package com.nec.cache
 
-import com.nec.spark.planning.VeColColumnarVector.{CachedColVector, DualVeBatch}
+import com.nec.cache.VeColColumnarVector.CachedColumnVector
 import com.nec.ve.VeColBatch
 import org.apache.spark.sql.columnar.CachedBatch
 
 object CachedVeBatch {
-  def apply(ccv: List[CachedColVector]): CachedVeBatch = CachedVeBatch(DualVeBatch(ccv))
+  def apply(ccv: List[CachedColumnVector]): CachedVeBatch = CachedVeBatch(DualColumnarBatchContainer(ccv))
   def apply(veColBatch: VeColBatch): CachedVeBatch = CachedVeBatch(
-    DualVeBatch(veColBatch.cols.map(vcv => Left(vcv)))
+    DualColumnarBatchContainer(veColBatch.cols.map(vcv => Left(vcv)))
   )
 }
-final case class CachedVeBatch(dualVeBatch: DualVeBatch) extends CachedBatch {
+final case class CachedVeBatch(dualVeBatch: DualColumnarBatchContainer) extends CachedBatch {
   override def numRows: Int = dualVeBatch.numRows
 
   override def sizeInBytes: Long = dualVeBatch.onCpuSize.getOrElse {
