+++ src/test/scala/org/apache/spark/sql/vectorized/DualModeTest.scala
@@ -1,7 +1,7 @@
 package org.apache.spark.sql.vectorized
 
+import com.nec.cache.VeColColumnarVector
 import com.nec.spark.agile.CFunctionGeneration.{VeScalarType, VeType}
-import com.nec.spark.planning.VeColColumnarVector
 import com.nec.ve.VeColBatch
 import com.nec.ve.VeColBatch.{VeColVector, VeColVectorSource}
 import org.apache.spark.sql.catalyst.InternalRow
